refactor: WriteCommentArea 리팩토링
