+++ src/components/Feeds/Feed/FeedItem/CommentsArea/WriteCommentArea/CommentInput.styled.js
@@ -1,0 +1,8 @@
+import { TextInput } from 'components/common';
+import styled from 'styled-components';
+
+const CommentTextInput = styled(TextInput)`
+  background-color: ${({ theme: { color } }) => color.white};
+`;
+
+export default CommentTextInput;
