+++ src/components/Feeds/Feed/FeedItem/CommentsArea/WriteCommentArea/WriteCommentArea.jsx
@@ -3,13 +3,12 @@ import LOCAL_STORAGE_KEY_NAME from 'constants';
 import { useState, useEffect } from 'react';
 
 import { Icon } from 'components/common';
-import theme from 'styles/theme';
 import storageUtils from 'utils/storage';
 import isEmpty from 'utils/utils';
 
+import CommentInput from './CommentInput';
 import {
   WriteCommentForm,
-  CommentTextInput,
   CommentSubmitButton,
 } from './WriteCommentArea.styled';
 
@@ -17,22 +16,6 @@ const COMMENT_INPUT_NAME = 'comment-content';
 const EMPTY_COMMENT_ALRET_MESSAGE = '내용을 입력해주세요.';
 
 const { getLocalStorage } = storageUtils;
-
-const CommentInputText = ({ inputText, setInputText }) => {
-  const handleOnChange = ({ target }) => {
-    const inputValue = target.value;
-    setInputText(inputValue);
-  };
-
-  return (
-    <CommentTextInput
-      size={{ width: theme.comment.inputWidth, height: '100%' }}
-      name={COMMENT_INPUT_NAME}
-      inputValue={inputText}
-      onChange={handleOnChange}
-    />
-  );
-};
 
 const WriteCommentArea = ({ comments, setComments }) => {
   const [inputText, setInputText] = useState('');
@@ -65,7 +48,11 @@ const WriteCommentArea = ({ comments, setComments }) => {
   return (
     <WriteCommentForm onSubmit={onSubmitHandler}>
       <Icon icon="smile" />
-      <CommentInputText inputText={inputText} setInputText={setInputText} />
+      <CommentInput
+        inputText={inputText}
+        setInputText={setInputText}
+        inputName={COMMENT_INPUT_NAME}
+      />
       <CommentSubmitButton type="submit">게시</CommentSubmitButton>
     </WriteCommentForm>
   );
