+++ src/components/Feeds/Feed/FeedItem/CommentsArea/WriteCommentArea/index.js
@@ -1,0 +1,3 @@
+import WriteCommentArea from './WriteCommentArea';
+
+export default WriteCommentArea;
